feat(DTFS2-8122): moved to helper
