+++ libs/common/src/helpers/index.ts
@@ -55,3 +55,4 @@ export * from './get-maintenance-timestamp';
 export * from './is-https';
 export * from './configure-express-application';
 export * from './configure-redis-cache';
+export * from './configure-express-session';
