+++ libs/common/src/middleware/index.ts
@@ -3,4 +3,3 @@ export * from './create-validation-middleware-for-schema';
 export * from './set-no-store-cache-control';
 export * from './validate-mongoId';
 export * from './maintainance';
-export * from './session';
